Updated collision on ships -ships willnot bounce of anything rather go through stuff. -only when the actor ships overlap with will do damage or kill said ship will react
We need to figure out how to get some sort of filter for diffrent type of ship events, like bouncing the ship or blowing it up that will run when the ship overlaps with something. currently
using HIt or Overlap events isn't working when trying to compare actors it overlaps with with spesific actors for results.